backend/pyproject.toml
[tool.poetry]
name = "stock-social-sentiment-backend"
version = "0.1.0"
description = "Backend for stock social sentiment analysis"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
alembic = "^1.12.0"

# Scraper service dependencies
[tool.poetry.group.scraper]
optional = true

[tool.poetry.group.scraper.dependencies]
tweepy = "^4.14.0"
praw = "^7.7.1"

# Processor service dependencies
[tool.poetry.group.processor]
optional = true

[tool.poetry.group.processor.dependencies]
nltk = "^3.8.1"
transformers = "^4.31.0"
torch = "2.2.2"
pandas = "^2.1.0"

# API service dependencies
[tool.poetry.group.api]
optional = true

[tool.poetry.group.api.dependencies]
fastapi = "^0.109.0"
uvicorn = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 

